data_preprocessing/requirements.txt
numpy==1.21.2
python-dotenv==0.19.0
scipy==1.7.0

